Fixed issues.
