+++ BenchBnB/Models/DTOs/BenchDTO.cs
@@ -9,6 +9,8 @@ namespace BenchBnB.Models.DTOs
         public int Seats { get; set; }
         public string Description { get; set; }
         public decimal? AverageRating { get; set; }
+        public decimal Latitude { get; set; }
+        public decimal Longitude { get; set; }
 
         public User User { get; set; }
