+++ BenchBnB/Scripts/BenchIndex/clickableBenchList.js
@@ -6,8 +6,5 @@ table.addEventListener('click', (event) => {
     const tr = event.target.parentNode;
     const redirectLink = tr.getAttribute('data')
 
-    console.log(tr);
-    console.log(redirectLink);
-
     window.location.href = redirectLink;
 });
