+++ BenchBnB/Scripts/BenchIndex/tableGenerator.js
@@ -24,26 +24,27 @@
             }
             if (minSeats.value !== '' && maxSeats.value === ''){
                 benchInfo.forEach(bench => {
-                    if (bench.Seats >= minSeats.value) {
+                    if (bench.Seats >= parseInt(minSeats.value,10)) {
                         filteredBenches.push(bench);
                     }
                 });
             }
             if (minSeats.value === '' && maxSeats.value !== ''){          
                 benchInfo.forEach(bench => {
-                    if (bench.Seats <= maxSeats.value) {
+                    if (bench.Seats <= parseInt(maxSeats.value,10)) {
                         filteredBenches.push(bench);
                     }
                 });
             }
             if (minSeats.value !== '' && maxSeats.value !== ''){ 
                 benchInfo.forEach(bench => {
-                    if (bench.Seats >= minSeats.value && bench.Seats <= maxSeats.value) {
+                    if (bench.Seats >= parseInt(minSeats.value,10) && bench.Seats <= parseInt(maxSeats.value,10)) {
                         filteredBenches.push(bench);
                     }
                 });
             }
             createTable(filteredBenches);
+            generateMap(filteredBenches);
         });
     };
 
@@ -98,8 +99,61 @@
         myTableDiv.innerHTML = HTML;
     };
 
+    const generateMap = function (benchList) {
+        var map = new ol.Map({
+            target: 'map',
+            layers: [
+                new ol.layer.Tile({
+                    source: new ol.source.OSM()
+                })
+            ],
+            view: new ol.View({
+                center: ol.proj.fromLonLat([-73.145, 40.7891]),
+                zoom: 8
+            })
+        });
+
+        var vectorSource = new ol.source.Vector({});
+
+        const markerArray = [];
+        benchList.forEach(bench => {
+            var marker = new ol.Feature({
+                geometry: new ol.geom.Point(
+                    ol.proj.fromLonLat([bench.Longitude, bench.Latitude])
+                )
+            });
+            vectorSource.addFeature(marker);
+        });
+
+        var markerVectorLayer = new ol.layer.Vector({
+            source: vectorSource
+        });
+        map.addLayer(markerVectorLayer);
+
+        map.on('singleclick', async function (event) {
+            event.coordinate = ol.proj.transform(event.coordinate, 'EPSG:3857', 'EPSG:4326');
+            latitude = event.coordinate[1];
+            longitude = event.coordinate[0];
+
+            let alreadyExists = false;
+
+            benchInfo.forEach(bench => {
+                if (Math.abs(bench.Latitude - latitude) < 0.00025 && Math.abs(bench.Longitude - longitude) < 0.00025){
+                    window.location.href = "Bench/View/" + bench.Id;
+                    alreadyExists = true;
+                }
+            });
+            if (!alreadyExists)
+            {
+                window.location.href = "Bench/Add?Lat=" + latitude + "&Lon=" + longitude;
+            }
+        });
+    }
+
     // Default Full Table
     createTable(benchInfo);
 
     filterMinMaxSeats();
+
+    generateMap(benchInfo);
 })();
