+++ BenchBnB/Views/Bench/Index.cshtml
@@ -25,9 +25,9 @@
     
 </div>
 
+<br />
+<p>@Html.ActionLink("Add A Bench!", "Add", null, new { @class = "btn btn-success" })</p>
 
 <script src="~/Scripts/BenchIndex/tableGenerator.js"></script>
 <script src="~/Scripts/BenchIndex/map.js"></script>
 <script src="~/Scripts/BenchIndex/clickableBenchList.js"></script>
-<br />
-<p>@Html.ActionLink("Add A Bench!", "Add", null, new { @class = "btn btn-success" })</p>
